enhanced Calculator tests

+++ src/main/java/com/myprojects/junittests/Calculator.java
@@ -14,12 +14,12 @@ public class Calculator {
         return a * b;
     }
 
-    public float divide(Integer a, Integer b) {
+    public double divide(double a, double b) {
         if (b == 0) {
             System.out.println("Division par 0 impossible !");
             return 0;
         }
-        float res = a / b;
+        double res = a / b;
         return res;
     }
 }
